# PERF: MultiIndex.argsort / MultiIndex.sort_values