--- a/asv_bench/benchmarks/multiindex_object.py
+++ b/asv_bench/benchmarks/multiindex_object.py
@@ -8,6 +8,7 @@ from pandas import (
     MultiIndex,
     RangeIndex,
     Series,
+    array,
     date_range,
 )
 
@@ -177,12 +178,16 @@ class Sortlevel:
 
 
 class SortValues:
-    def setup(self):
-        a = np.tile(np.arange(100), 1000)
-        b = np.tile(np.arange(1000), 100)
+
+    params = ["int64", "Int64"]
+    param_names = ["dtype"]
+
+    def setup(self, dtype):
+        a = array(np.tile(np.arange(100), 1000), dtype=dtype)
+        b = array(np.tile(np.arange(1000), 100), dtype=dtype)
         self.mi = MultiIndex.from_arrays([a, b])
 
-    def time_sort_values(self):
+    def time_sort_values(self, dtype):
         self.mi.sort_values()