- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v1.6.0.rst` file if fixing a bug or adding a new feature.

Perf improvement in `MultiIndex.argsort` and `MultiIndex.sort_values`.

Also, an xfail was removed because np.lexsort does not complain about complex dtype arrays whereas np.argsort will complain about an object array of tuples containing complex and non-complex dtypes. 

ASV added:

```
from asv_bench.benchmarks.multiindex_object import SortValues

b = SortValues()
b.setup()

%timeit b.time_sort_values()
```


```
88.6 ms ± 10 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)    <- main
2 ms ± 30.3 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)  <- PR
```