@lukemanley 2022-09-06
> Could you add a Benchmark for Nullable dtypes too?

Added a benchmark for nullable Int64 and updated the timings in the summary. However, the nullable index used in the asv does not contain any pd.NA values as numpy will raise if it encounters pd.NA. That is true for numpy.argsort (with a list of tuples) and numpy.lexsort so this isn't new behavior. Can handling pd.NA be an enhancement/bug fix for a separate PR?